api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
PyJWT==2.8.0
bcrypt==4.1.3
aiohttp==3.9.5
stripe==9.5.0
mercadopago==2.3.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
aioresponses==0.7.6
aiosqlite==0.20.0
